# openapi.yaml
openapi: 3.1.0
info:
  title: Henry M2 - API Miniblog
  version: 1.0.0
  description: API REST para gestión de autores y posts. Proyecto Integrador M2.
  contact:
    name: Santiago Ezequiel Chavez
    email: santiago@example.com # Cambialo por el tuyo
servers:
  - url: http://localhost:3000
    description: Servidor de desarrollo local
  - url: https://api-miniblog-production.up.railway.app # Cambialo cuando tengas la de Railway
    description: Servidor de producción en Railway

paths:
  /api/authors:
    get:
      summary: Obtener todos los autores
      responses:
        "200":
          description: Lista de autores devuelta con éxito.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Author'
    post:
      summary: Crear un nuevo autor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewAuthor'
      responses:
        "201":
          description: Autor creado exitosamente.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Author'
        "400":
          description: Datos inválidos o email duplicado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/authors/{id}:
    get:
      summary: Obtener autor por ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Autor encontrado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Author'
        "404":
          description: Autor no encontrado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      summary: Actualizar autor por ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewAuthor'
      responses:
        "200":
          description: Autor actualizado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Author'
        "400":
          description: Datos inválidos o email duplicado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: Autor no encontrado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      summary: Eliminar autor por ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Autor eliminado correctamente.
        "404":
          description: Autor no encontrado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/posts:
    get:
      summary: Obtener todos los posts
      responses:
        "200":
          description: Lista de posts devuelta con éxito.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Post'
    post:
      summary: Crear un nuevo post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewPost'
      responses:
        "201":
          description: Post creado exitosamente.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        "400":
          description: Datos inválidos.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: Error interno al crear post (por ejemplo, author_id inválido).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/posts/{id}:
    get:
      summary: Obtener post por ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Post encontrado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        "404":
          description: Post no encontrado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      summary: Actualizar post por ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, content]
              properties:
                title:
                  type: string
                content:
                  type: string
                published:
                  type: boolean
      responses:
        "200":
          description: Post actualizado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        "400":
          description: Datos inválidos.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: Post no encontrado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      summary: Eliminar post por ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Post eliminado correctamente.
        "404":
          description: Post no encontrado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/posts/author/{authorId}:
    get:
      summary: Obtener posts de un autor específico
      parameters:
        - name: authorId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lista de posts del autor.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Post'
components:
  schemas:
    Author:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
          format: email
        bio:
          type: string
    NewAuthor:
      type: object
      required:
        - name
        - email
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        bio:
          type: string
    Post:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        content:
          type: string
        author_id:
          type: integer
        published:
          type: boolean
    NewPost:
      type: object
      required:
        - title
        - content
        - author_id
      properties:
        title:
          type: string
        content:
          type: string
        author_id:
          type: integer
        published:
          type: boolean
          default: false
    ErrorResponse:
      type: object
      properties:
        error:
          type: string